requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Django==4.2
django-extensions==3.2.3
django-jalali-date==1.1.0
django-model-utils==4.5.0
django-render-partial==0.4
executing==2.0.1
idna==3.7
ipython==8.22.2
isodate==0.6.1
jalali_core==1.0.0
jdatetime==5.0.0
jedi==0.19.1
lxml==5.2.1
matplotlib-inline==0.1.6
parso==0.8.3
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
setuptools==69.2.0
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.43.0
zeep==4.2.1
